requirements.txt
tiktoken
qdrant-client
langchain-community
openai
PyPDF2
huggingface
python-dotenv